+++ common/src/main/java/uk/me/desert_island/rer/RerDataComponents.java
@@ -9,9 +9,13 @@ import net.minecraft.util.ExtraCodecs;
 import java.util.function.UnaryOperator;
 
 public class RerDataComponents {
-    public static DataComponentType<Integer> RER_COUNT = register("RER_COUNT", builder -> builder.persistent(ExtraCodecs.intRange(1, 99)).networkSynchronized(ByteBufCodecs.VAR_INT));
+    public static DataComponentType<Integer> RER_COUNT = register("rer:rer_count", builder -> builder.persistent(ExtraCodecs.intRange(1, 99)).networkSynchronized(ByteBufCodecs.VAR_INT));
 
     private static <T> DataComponentType<T> register(String string, UnaryOperator<DataComponentType.Builder<T>> unaryOperator) {
         return Registry.register(BuiltInRegistries.DATA_COMPONENT_TYPE, string, unaryOperator.apply(DataComponentType.builder()).build());
+    }
+
+    public static void init() {
+        // :)
     }
 }
